Build delete confirmation for landlord and rentholder fixed some bugs.

+++ src/app/components/landlord-components/rent-holder/rent-holder.component.ts
@@ -9,7 +9,7 @@ import {ToastrService} from 'ngx-toastr';
   styleUrls: ['./rent-holder.component.css']
 })
 export class RentHolderComponent {
-  constructor(private landlordServ:LandlordService,private spinner:NgxSpinnerService,private toster:ToastrService){}
+  constructor(private toaster:ToastrService,private landlordServ:LandlordService,private spinner:NgxSpinnerService,private toster:ToastrService){}
   users:any=[];
 
 ngOnInit(){
@@ -37,19 +37,25 @@ downloadDoc(link:any){
   document.body.removeChild(downloadLink);
 }
 deleteRentHolder(id:any){
-  this.landlordServ.deleteRentHolderData(id).subscribe({
-    next:(res:any)=>{
-      if(res.status){
-        this.toster.success('Rent-Holder Deleted.','Success');
-        this.ngOnInit();
+  let confString = prompt("This Action is irreversible and permanently delete all releted data, To confirm please type ['delete rentholder'] bellow.");
+  
+  if(confString ==='delete rentholder'){
+    this.landlordServ.deleteRentHolderData(id).subscribe({
+      next:(res:any)=>{
+        if(res.status){
+          this.toaster.success(`Rent Holder Deleted`,'Success');
+          this.ngOnInit();
+        }
+      },error:(error)=>{
+        console.error(error);
+        this.toaster.error(`Something wents wrong.`);
       }
-    },
-    error:(err)=>{
-      console.log(err.error);
-      this.toster.error('Something went wrong.','Error');
-      
+    });
+  }else if(confString===null){
+  }else{
+    this.toaster.error("Sorry! wrong command.");
+  }
 
-    }
-  })
+  
 }
 }
